ADDED AI! Senses, Brains, Memory!
A pretty big update, but largely AI. Added a brain class that coordinates
the current state of the creature with it's senses. Made creatures Finite
State Machines with a active and inactive states, active when they have
recently seen the player, inactive otherwise. Added senses (currently just
sight) which tell the creature where the player is.